// client/src/ticket-api-types.ts
export interface Ticket {
  /** @format uuid */
  id?: string;
  title?: string | null;
  body?: string | null;

  /** @format date-time */
  createdDate?: string;

  /** @format date-time */
  lastUpdatedDate?: string;
  requesterId?: string | null;
  requester?: User;
  comments?: TicketComment[] | null;
}

export interface TicketComment {
  /** @format uuid */
  id?: string;
  comment?: string | null;

  /** @format date-time */
  createdDate?: string;

  /** @format date-time */
  lastUpdatedDate?: string;
  author?: User;
  ticketCommentType?: TicketCommentType;
}

export interface TicketCommentType {
  /** @format uuid */
  ticketCommentTypeId?: string;
  name?: string | null;
}

export interface User {
  id?: string | null;
  userName?: string | null;
  tickets?: Ticket[] | null;
  comments?: TicketComment[] | null;
}